Support multiple notebook controllers and Python environments (#26)
Previously marimo-lsp (and in turn the VS Code extension) spawned marimo
kernel from the _same_ Python environment running the language server.
This meant there is only one conceptual "notebook controller" in the
eyes of VS Code, and has several problems making usability of VS Code
extension with this prior architecture impractical:

1.) There could only be one Python version
2.) There could only be one Python environment
3.) All notebooks inside VS Code had to share same environment
4.) Installing deps into this shared environment "pollutes" the parent

With the changes in:

- https://github.com/marimo-team/marimo/pull/6303
- https://github.com/marimo-team/marimo/pull/6234

We now have more generic messaging and most importantly the ability to
separate _kernels_ from the parent process which "launches" the kernel.
This PR updates the extension with the ability to discover existing
Python environments and configure the kernel launcher to "launch" from
these environments when running cells. Effectively, users can "select" a
Python environment for which to "run" the current notebook (and have
those environments be isolated and separate from each other).