Support missing package installation with `uv` (#64)
Add a custom handler for `missing-package-alert` from the kernel,
prompting the user to "Install All" or "Customize..." which packages to
install via a text input with `uv`.

This currently bypasses the package manger built into marimo and
supports `uv` only. Ideally we could unify the experience to piggyback
on what we've implemented for more package managers inside marimo for
more package managers, but the separate of kernel and kernel manager is
causing issues where the package installer is trying to install into the
`marimo-lsp` environment (not the individual kernels).