Harden `cell.metadata` schema and use `cell.languageId` as source of truth
These changes strengthens how VS Code cell metadata is defined and
validated. It also removes the `language` field from cell metadata, in
preference of using `cell.languageId` as the source of truth for the
"current" state of the cell.

Previously, the serializer would write this value, but it could easily
drift from the actual cell language in the VS Code UI. The untyped
nature of `cell.metadata` lead to hidden any's that didn't help us spot
this.