Require naming new marimo notebook
Untitled notebooks lead to some awkwardness for users. Since we are pure
Python, saving from `untitled://` -> `file://` unloads the notebook view
even if the user has opted for it. 

This new flow requires the user to name a notebook first, which we
create and save to disk and then display with our renderer.