Provide code-lens to Open as a notebook (#261)
This is the same as the `Open as notebook` in the top right, but can be
another entry point.

Open to not including this if its duplicate UI for no reason. However,
in Cursor, the top-right marimo icon did go away (not sure what setting
forced that or if it was a window size issue).


<img width="262" height="208" alt="image"
src="https://github.com/user-attachments/assets/08ec507e-8d9f-4a80-8a2d-d15b73b0697f"
/>