Add support for auto-reload config (#260)
This adds auto-reload config to the notebook navbar and DRYs up some
similar code for toggling on-cell-change. This works well now with the
recent autoreload fixes/changes in the past PRs.

I need to create an command/action for each toggle state, because vscode
does not allow for dynamic navbar titles. This works fine, but does
clutter up the command palette with each option.

Closes #254

---------

Co-authored-by: pre-commit-ci[bot] <66853113+pre-commit-ci[bot]@users.noreply.github.com>